Fix some C4244 warnings (x64 specific).
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@5011 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/mpc-hc/FGManager.cpp
@@ -1132,7 +1132,7 @@ STDMETHODIMP CFGManager::ConnectFilter(IBaseFilter* pBF, IPin* pPinIn)
 			HRESULT hr = Connect(pPin, pPinIn);
 
 			if (SUCCEEDED(hr)) {
-				for (int i = m_deadends.GetCount()-1; i >= 0; i--)
+				for (ptrdiff_t i = m_deadends.GetCount()-1; i >= 0; i--)
 					if (m_deadends[i]->Compare(m_streampath)) {
 						m_deadends.RemoveAt(i);
 					}
